feat: consolidate book fetching hooks and remove unused ones
